@@ -13,20 +13,32 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+
 from typing import Optional
 
 from aiohttp import ClientSession
 
 from monday_async import __version__
 from monday_async.resources import (
-    APIResource, CustomResource, WebhooksResource, NotificationResource, UsersResource, WorkspaceResource,
-    FolderResource, BoardResource, TagResource, ColumnResource, GroupResource, ItemResource, UpdateResource,
-    ComplexityResource, AccountResource, TeamsResource
+    AccountResource,
+    APIResource,
+    BoardResource,
+    ColumnResource,
+    ComplexityResource,
+    CustomResource,
+    FolderResource,
+    GroupResource,
+    ItemResource,
+    NotificationResource,
+    TagResource,
+    TeamsResource,
+    UpdateResource,
+    UsersResource,
+    WebhooksResource,
+    WorkspaceResource,
 )
 
-_DEFAULT_HEADERS = {
-    "API-Version": "2025-01"
-}
+_DEFAULT_API_VERSION = "2025-04"
 
 
 class AsyncMondayClient:
@@ -50,19 +62,29 @@ class AsyncMondayClient:
         updates (UpdateResource):
     """
 
-    def __init__(self, token: str, session: Optional[ClientSession] = None, headers: dict = None):
+    def __init__(
+        self,
+        token: str,
+        session: ClientSession | None = None,
+        headers: Optional[dict] = None,
+        api_version: str = _DEFAULT_API_VERSION,
+    ):
         """
         Args:
             token (str): Your monday.com API access token.
             session (ClientSession): Optional, externally managed aiohttp session. Recommended to use the same session
                 for all the requests.
             headers (dict): Additional headers to send with each request.
+            api_version (str): monday.com API version to use. Defaults to "2025-04"
         """
         self._session = session
         self._external_session = True if session else False
+        self.api_version = api_version
 
         if not headers:
-            headers = _DEFAULT_HEADERS.copy()
+            headers = {"API-Version": api_version}
+        elif "API-Version" not in headers:
+            headers["API-Version"] = api_version
 
         self.complexity = ComplexityResource(token=token, headers=headers, session=self._session)
         self.custom = CustomResource(token=token, headers=headers, session=self._session)
@@ -82,7 +104,7 @@ class AsyncMondayClient:
         self.updates = UpdateResource(token=token, headers=headers, session=self._session)
 
     def __enter__(self):
-        raise RuntimeError('Use `async with AsyncMondayClient(...)` instead of `with AsyncMondayClient(...)`')
+        raise RuntimeError("Use `async with AsyncMondayClient(...)` instead of `with AsyncMondayClient(...)`")
 
     def __exit__(self, exc_type, exc, tb):
         pass
@@ -98,7 +120,7 @@ class AsyncMondayClient:
             await self._session.close()
 
     def __str__(self):
-        return f'AsyncMondayClient {__version__}'
+        return f"AsyncMondayClient {__version__}"
 
     def __repr__(self):
-        return f'AsyncMondayClient {__version__}'
+        return f"AsyncMondayClient {__version__}"
