@@ -13,30 +13,40 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import List, Union, Optional
+from typing import Optional, Union
 
-from monday_async.resources.base_resource import AsyncBaseResource
-from monday_async.types import (BoardKind, State, SubscriberKind, BoardAttributes, DuplicateBoardType,
-                                BoardsOrderBy)
-from monday_async.utils.queries import (
-    get_boards_query, create_board_query, duplicate_board_query,
-    update_board_query, archive_board_query, delete_board_query,
-    add_users_to_board_query, remove_users_from_board_query,
-    add_teams_to_board_query, delete_teams_from_board_query,
-    get_board_views_query
+from monday_async.graphql.mutations import (
+    add_teams_to_board_mutation,
+    add_users_to_board_mutation,
+    archive_board_mutation,
+    create_board_mutation,
+    delete_board_mutation,
+    delete_teams_from_board_mutation,
+    duplicate_board_mutation,
+    remove_users_from_board_mutation,
+    update_board_mutation,
 )
+from monday_async.graphql.queries import get_board_views_query, get_boards_query
+from monday_async.resources.base_resource import AsyncBaseResource
+from monday_async.types import BoardAttributes, BoardKind, BoardsOrderBy, DuplicateBoardType, State, SubscriberKind
 
 ID = Union[int, str]
 
 
 class BoardResource(AsyncBaseResource):
-    async def get_boards(self, ids: Union[ID, List[ID]] = None,
-                         board_kind: Optional[BoardKind] = None,
-                         state: State = State.ACTIVE,
-                         workspace_ids: Union[ID, List[ID]] = None,
-                         order_by: Optional[BoardsOrderBy] = None,
-                         limit: int = 25, page: int = 1,
-                         with_columns: bool = True, with_groups: bool = True, with_complexity: bool = False) -> dict:
+    async def get_boards(
+        self,
+        ids: ID | list[ID] = None,
+        board_kind: BoardKind | None = None,
+        state: State = State.ACTIVE,
+        workspace_ids: ID | list[ID] = None,
+        order_by: BoardsOrderBy | None = None,
+        limit: int = 25,
+        page: int = 1,
+        with_columns: bool = True,
+        with_groups: bool = True,
+        with_complexity: bool = False,
+    ) -> dict:
         """
         Execute a query to retrieve boards, offering filtering by IDs, board kind, state, workspace,
          and ordering options.
@@ -55,21 +65,39 @@ class BoardResource(AsyncBaseResource):
             with_groups (bool): (Optional) Set to True to include groups in the query results.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = get_boards_query(ids=ids, board_kind=board_kind, state=state, workspace_ids=workspace_ids,
-                                 order_by=order_by, limit=limit, page=page, with_columns=with_columns,
-                                 with_groups=with_groups, with_complexity=with_complexity)
+        query = get_boards_query(
+            ids=ids,
+            board_kind=board_kind,
+            state=state,
+            workspace_ids=workspace_ids,
+            order_by=order_by,
+            limit=limit,
+            page=page,
+            with_columns=with_columns,
+            with_groups=with_groups,
+            with_complexity=with_complexity,
+        )
         return await self.client.execute(query)
 
-    async def create_board(self, board_name: str, board_kind: BoardKind,
-                           description: Optional[str] = None, folder_id: Optional[ID] = None,
-                           workspace_id: Optional[ID] = None,
-                           template_id: Optional[ID] = None, board_owner_ids: List[ID] = None,
-                           board_owner_team_ids: List[ID] = None, board_subscriber_ids: List[ID] = None,
-                           board_subscriber_teams_ids: List[ID] = None, empty: bool = False,
-                           with_columns: bool = False, with_groups: bool = False,
-                           with_complexity: bool = False) -> dict:
+    async def create_board(
+        self,
+        board_name: str,
+        board_kind: BoardKind,
+        description: str | None = None,
+        folder_id: ID | None = None,
+        workspace_id: ID | None = None,
+        template_id: ID | None = None,
+        board_owner_ids: Optional[list[ID]] = None,
+        board_owner_team_ids: Optional[list[ID]] = None,
+        board_subscriber_ids: Optional[list[ID]] = None,
+        board_subscriber_teams_ids: Optional[list[ID]] = None,
+        empty: bool = False,
+        with_columns: bool = False,
+        with_groups: bool = False,
+        with_complexity: bool = False,
+    ) -> dict:
         """
-        Execute a query to create a new board with specified name, kind, and optional description, folder, workspace,
+        Execute a mutation to create a new board with specified name, kind, and optional description, folder, workspace,
          template, and subscribers/owners.
 
         For more information, visit https://developer.monday.com/api-reference/reference/boards#create-a-board
@@ -90,22 +118,38 @@ class BoardResource(AsyncBaseResource):
             with_groups (bool): (Optional) Set to True to include groups in the query results. Defaults to False.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = create_board_query(board_name=board_name, board_kind=board_kind, description=description,
-                                   folder_id=folder_id, workspace_id=workspace_id, template_id=template_id,
-                                   board_owner_ids=board_owner_ids, board_owner_team_ids=board_owner_team_ids,
-                                   board_subscriber_ids=board_subscriber_ids,
-                                   board_subscriber_teams_ids=board_subscriber_teams_ids,
-                                   empty=empty, with_columns=with_columns,
-                                   with_groups=with_groups, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = create_board_mutation(
+            board_name=board_name,
+            board_kind=board_kind,
+            description=description,
+            folder_id=folder_id,
+            workspace_id=workspace_id,
+            template_id=template_id,
+            board_owner_ids=board_owner_ids,
+            board_owner_team_ids=board_owner_team_ids,
+            board_subscriber_ids=board_subscriber_ids,
+            board_subscriber_teams_ids=board_subscriber_teams_ids,
+            empty=empty,
+            with_columns=with_columns,
+            with_groups=with_groups,
+            with_complexity=with_complexity,
+        )
+        return await self.client.execute(mutation)
 
-    async def duplicate_board(self, board_id: ID, duplicate_type: DuplicateBoardType,
-                              board_name: Optional[str] = None, workspace_id: Optional[ID] = None,
-                              folder_id: Optional[ID] = None, keep_subscribers: bool = False,
-                              with_columns: bool = False, with_groups: bool = False,
-                              with_complexity: bool = False) -> dict:
+    async def duplicate_board(
+        self,
+        board_id: ID,
+        duplicate_type: DuplicateBoardType,
+        board_name: str | None = None,
+        workspace_id: ID | None = None,
+        folder_id: ID | None = None,
+        keep_subscribers: bool = False,
+        with_columns: bool = False,
+        with_groups: bool = False,
+        with_complexity: bool = False,
+    ) -> dict:
         """
-        Execute a query to duplicate a board with options to include structure, items, updates, and subscribers.
+        Execute a mutation to duplicate a board with options to include structure, items, updates, and subscribers.
 
         For more information, visit https://developer.monday.com/api-reference/reference/boards#duplicate-a-board
 
@@ -124,17 +168,24 @@ class BoardResource(AsyncBaseResource):
             with_groups (bool): (Optional) Set to True to include groups in the query results. Defaults to False.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = duplicate_board_query(board_id=board_id, duplicate_type=duplicate_type,
-                                      board_name=board_name, workspace_id=workspace_id,
-                                      folder_id=folder_id, keep_subscribers=keep_subscribers,
-                                      with_columns=with_columns, with_groups=with_groups,
-                                      with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = duplicate_board_mutation(
+            board_id=board_id,
+            duplicate_type=duplicate_type,
+            board_name=board_name,
+            workspace_id=workspace_id,
+            folder_id=folder_id,
+            keep_subscribers=keep_subscribers,
+            with_columns=with_columns,
+            with_groups=with_groups,
+            with_complexity=with_complexity,
+        )
+        return await self.client.execute(mutation)
 
-    async def update_board(self, board_id: ID, board_attribute: BoardAttributes, new_value: str,
-                           with_complexity: bool = False) -> dict:
+    async def update_board(
+        self, board_id: ID, board_attribute: BoardAttributes, new_value: str, with_complexity: bool = False
+    ) -> dict:
         """
-        Execute a query to update a board attribute.
+        Execute a mutation to update a board attribute.
 
         For more information, visit https://developer.monday.com/api-reference/reference/boards#update-a-board
 
@@ -144,13 +195,14 @@ class BoardResource(AsyncBaseResource):
             new_value (str): The new attribute value
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = update_board_query(board_id=board_id, board_attribute=board_attribute, new_value=new_value,
-                                   with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = update_board_mutation(
+            board_id=board_id, board_attribute=board_attribute, new_value=new_value, with_complexity=with_complexity
+        )
+        return await self.client.execute(mutation)
 
     async def archive_board(self, board_id: ID, with_complexity: bool = False) -> dict:
         """
-        Execute a query to archive a board, making it no longer visible in the active board list.
+        Execute a mutation to archive a board, making it no longer visible in the active board list.
 
         For more information, visit https://developer.monday.com/api-reference/reference/boards#archive-a-board
 
@@ -158,12 +210,12 @@ class BoardResource(AsyncBaseResource):
             board_id (ID): The ID of the board to archive.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = archive_board_query(board_id=board_id, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = archive_board_mutation(board_id=board_id, with_complexity=with_complexity)
+        return await self.client.execute(mutation)
 
     async def delete_board(self, board_id: ID, with_complexity: bool = False) -> dict:
         """
-        Execute a query to permanently delete a board.
+        Execute a mutation to permanently delete a board.
 
         For more information, visit https://developer.monday.com/api-reference/reference/boards#delete-a-board
 
@@ -171,13 +223,14 @@ class BoardResource(AsyncBaseResource):
             board_id (ID): The ID of the board to delete.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = delete_board_query(board_id=board_id, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = delete_board_mutation(board_id=board_id, with_complexity=with_complexity)
+        return await self.client.execute(mutation)
 
-    async def add_users_to_board(self, board_id: ID, user_ids: Union[ID, List[ID]],
-                                 kind: SubscriberKind, with_complexity: bool = False) -> str:
+    async def add_users_to_board(
+        self, board_id: ID, user_ids: ID | list[ID], kind: SubscriberKind, with_complexity: bool = False
+    ) -> str:
         """
-        Execute a query to add users as subscribers or owners to a board.
+        Execute a mutation to add users as subscribers or owners to a board.
 
         For more information, visit https://developer.monday.com/api-reference/reference/users#add-users-to-a-board
 
@@ -187,12 +240,14 @@ class BoardResource(AsyncBaseResource):
             kind (SubscriberKind): The type of subscription to grant: subscriber or owner.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = add_users_to_board_query(board_id=board_id, user_ids=user_ids, kind=kind,
-                                         with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = add_users_to_board_mutation(
+            board_id=board_id, user_ids=user_ids, kind=kind, with_complexity=with_complexity
+        )
+        return await self.client.execute(mutation)
 
-    async def remove_users_from_board(self, board_id: ID, user_ids: Union[ID, List[ID]],
-                                      with_complexity: bool = False) -> dict:
+    async def remove_users_from_board(
+        self, board_id: ID, user_ids: ID | list[ID], with_complexity: bool = False
+    ) -> dict:
         """
         Execute a query to remove users from a board's subscribers or owners.
 
@@ -204,11 +259,14 @@ class BoardResource(AsyncBaseResource):
             user_ids (Union[ID, List[ID]]): A list of user IDs to remove from the board.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = remove_users_from_board_query(board_id=board_id, user_ids=user_ids, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = remove_users_from_board_mutation(
+            board_id=board_id, user_ids=user_ids, with_complexity=with_complexity
+        )
+        return await self.client.execute(mutation)
 
-    async def add_teams_to_board(self, board_id: ID, team_ids: Union[ID, List[ID]],
-                                 kind: SubscriberKind, with_complexity: bool = False) -> dict:
+    async def add_teams_to_board(
+        self, board_id: ID, team_ids: ID | list[ID], kind: SubscriberKind, with_complexity: bool = False
+    ) -> dict:
         """
         Execute a query to add teams as subscribers or owners to a board.
 
@@ -220,12 +278,14 @@ class BoardResource(AsyncBaseResource):
             kind (SubscriberKind): The type of subscription to grant: subscriber or owner.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = add_teams_to_board_query(board_id=board_id, team_ids=team_ids, kind=kind,
-                                         with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = add_teams_to_board_mutation(
+            board_id=board_id, team_ids=team_ids, kind=kind, with_complexity=with_complexity
+        )
+        return await self.client.execute(mutation)
 
-    async def delete_teams_from_board(self, board_id: ID, team_ids: Union[ID, List[ID]],
-                                      with_complexity: bool = False) -> dict:
+    async def delete_teams_from_board(
+        self, board_id: ID, team_ids: ID | list[ID], with_complexity: bool = False
+    ) -> dict:
         """
         Execute a query to remove teams from a board's subscribers or owners.
 
@@ -236,11 +296,18 @@ class BoardResource(AsyncBaseResource):
             team_ids (Union[ID, List[ID]]): A list of team IDs to remove from the board.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = delete_teams_from_board_query(board_id=board_id, team_ids=team_ids, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = delete_teams_from_board_mutation(
+            board_id=board_id, team_ids=team_ids, with_complexity=with_complexity
+        )
+        return await self.client.execute(mutation)
 
-    async def get_board_views(self, board_id: ID, ids: Union[ID, List[ID]] = None,
-                              view_type: Optional[str] = None, with_complexity: bool = False) -> dict:
+    async def get_board_views(
+        self,
+        board_id: ID,
+        ids: ID | list[ID] = None,
+        view_type: str | None = None,
+        with_complexity: bool = False,
+    ) -> dict:
         """
         Execute a query to retrieve the views associated with a specific board.
 
