@@ -13,25 +13,32 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import List, Union, Optional
+from typing import Union
 
+from monday_async.graphql.mutations import (
+    change_column_description_mutation,
+    change_column_title_mutation,
+    create_column_mutation,
+    delete_column_mutation,
+)
+from monday_async.graphql.queries import get_columns_by_board_query
 from monday_async.resources.base_resource import AsyncBaseResource
 from monday_async.types import ColumnType
-from monday_async.utils.queries import (
-    get_columns_by_board_query, create_column_query,
-    change_column_title_query, change_column_description_query,
-    delete_column_query
-)
 
 ID = Union[int, str]
 
 
 class ColumnResource(AsyncBaseResource):
-    async def get_columns_by_board(self, board_id: ID, ids: Union[ID, List[ID]] = None,
-                                   types: Union[ColumnType, List[ColumnType]] = None,
-                                   with_complexity: bool = False) -> dict:
+    async def get_columns_by_board(
+        self,
+        board_id: ID,
+        ids: ID | list[ID] = None,
+        types: ColumnType | list[ColumnType] = None,
+        with_complexity: bool = False,
+    ) -> dict:
         """
-        Execute a query to retrieve columns associated with a specific board, allowing filtering by column IDs and types.
+        Execute a query to retrieve columns associated with a specific board, allowing filtering
+        by column IDs and types.
 
         For more information, visit https://developer.monday.com/api-reference/reference/columns#queries
 
@@ -44,12 +51,19 @@ class ColumnResource(AsyncBaseResource):
         query = get_columns_by_board_query(board_id=board_id, ids=ids, types=types, with_complexity=with_complexity)
         return await self.client.execute(query)
 
-    async def create_column(self, board_id: ID, title: str, column_type: ColumnType,
-                            description: Optional[str] = None, defaults: Optional[dict] = None,
-                            column_id: Optional[str] = None, after_column_id: Optional[ID] = None,
-                            with_complexity: bool = False) -> dict:
+    async def create_column(
+        self,
+        board_id: ID,
+        title: str,
+        column_type: ColumnType,
+        description: str | None = None,
+        defaults: dict | None = None,
+        column_id: str | None = None,
+        after_column_id: ID | None = None,
+        with_complexity: bool = False,
+    ) -> dict:
         """
-        Execute a query to create a new column on a specific board with a specified title, type,
+        Execute a mutation to create a new column on a specific board with a specified title, type,
         and optional description, defaults, user-specified ID, and positioning.
 
         For more information, visit https://developer.monday.com/api-reference/reference/columns#create-a-column
@@ -69,15 +83,23 @@ class ColumnResource(AsyncBaseResource):
             after_column_id (ID): (Optional) The ID of the column after which to insert the new column.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = create_column_query(board_id=board_id, title=title, column_type=column_type,
-                                    description=description, defaults=defaults, column_id=column_id,
-                                    after_column_id=after_column_id, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = create_column_mutation(
+            board_id=board_id,
+            title=title,
+            column_type=column_type,
+            description=description,
+            defaults=defaults,
+            column_id=column_id,
+            after_column_id=after_column_id,
+            with_complexity=with_complexity,
+        )
+        return await self.client.execute(mutation)
 
-    async def change_column_title(self, board_id: ID, column_id: str, title: str,
-                                  with_complexity: bool = False) -> dict:
+    async def change_column_title(
+        self, board_id: ID, column_id: str, title: str, with_complexity: bool = False
+    ) -> dict:
         """
-        Execute a query to update the title of an existing column on a specific board.
+        Execute a mutation to update the title of an existing column on a specific board.
 
         For more information, visit https://developer.monday.com/api-reference/reference/columns#change-a-column-title
 
@@ -87,14 +109,16 @@ class ColumnResource(AsyncBaseResource):
             title (str): The new title for the column.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = change_column_title_query(board_id=board_id, column_id=column_id, title=title,
-                                          with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = change_column_title_mutation(
+            board_id=board_id, column_id=column_id, title=title, with_complexity=with_complexity
+        )
+        return await self.client.execute(mutation)
 
-    async def change_column_description(self, board_id: ID, column_id: str, description: str,
-                                        with_complexity: bool = False) -> dict:
+    async def change_column_description(
+        self, board_id: ID, column_id: str, description: str, with_complexity: bool = False
+    ) -> dict:
         """
-        Execute a query to update the description of an existing column on a specific board.
+        Execute a mutation to update the description of an existing column on a specific board.
 
         For more information, visit https://developer.monday.com/api-reference/reference/columns#change-column-metadata
 
@@ -104,14 +128,14 @@ class ColumnResource(AsyncBaseResource):
             description (str): The new description for the column.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = change_column_description_query(board_id=board_id, column_id=column_id,
-                                                description=description, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = change_column_description_mutation(
+            board_id=board_id, column_id=column_id, description=description, with_complexity=with_complexity
+        )
+        return await self.client.execute(mutation)
 
-    async def delete_column(self, board_id: ID, column_id: str,
-                            with_complexity: bool = False) -> dict:
+    async def delete_column(self, board_id: ID, column_id: str, with_complexity: bool = False) -> dict:
         """
-        Execute a query to remove a column from a specific board.
+        Execute a mutation to remove a column from a specific board.
 
         For more information, visit https://developer.monday.com/api-reference/reference/columns#delete-a-column
 
@@ -120,5 +144,5 @@ class ColumnResource(AsyncBaseResource):
             column_id (str): The unique identifier of the column to delete.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = delete_column_query(board_id=board_id, column_id=column_id, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = delete_column_mutation(board_id=board_id, column_id=column_id, with_complexity=with_complexity)
+        return await self.client.execute(mutation)
