@@ -13,13 +13,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import List, Union, Optional
+from typing import Union
 
-from monday_async.resources.base_resource import AsyncBaseResource
-from monday_async.utils.queries import (
-    get_updates_query, create_update_query, like_update_query, delete_update_query, add_file_to_update,
-    pin_update_query, unpin_update_query, unlike_update_query, edit_update_query
+from monday_async.graphql.mutations import (
+    add_file_to_update_mutation,
+    create_update_mutation,
+    delete_update_mutation,
+    edit_update_mutation,
+    like_update_mutation,
+    pin_update_mutation,
+    unlike_update_mutation,
+    unpin_update_mutation,
 )
+from monday_async.graphql.queries import get_updates_query
+from monday_async.resources.base_resource import AsyncBaseResource
 
 ID = Union[int, str]
 
@@ -31,8 +38,14 @@ class UpdateResource(AsyncBaseResource):
     For many users, the updates section is their primary form of communication within the platform.
     """
 
-    async def get_updates(self, ids: Union[ID, List[ID]] = None, limit: int = 25, page: int = 1,
-                          with_viewers: bool = False, with_complexity: bool = False) -> dict:
+    async def get_updates(
+        self,
+        ids: ID | list[ID] = None,
+        limit: int = 25,
+        page: int = 1,
+        with_viewers: bool = False,
+        with_complexity: bool = False,
+    ) -> dict:
         """
         Execute a query to retrieve updates, allowing pagination and filtering by update IDs.
 
@@ -45,14 +58,16 @@ class UpdateResource(AsyncBaseResource):
             with_viewers (bool): Set to True to return the viewers of the update.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = get_updates_query(ids=ids, limit=limit, page=page, with_viewers=with_viewers,
-                                  with_complexity=with_complexity)
+        query = get_updates_query(
+            ids=ids, limit=limit, page=page, with_viewers=with_viewers, with_complexity=with_complexity
+        )
         return await self.client.execute(query)
 
-    async def create_update(self, body: str, item_id: ID, parent_id: Optional[ID] = None,
-                            with_complexity: bool = False) -> dict:
+    async def create_update(
+        self, body: str, item_id: ID, parent_id: ID | None = None, with_complexity: bool = False
+    ) -> dict:
         """
-        Execute a query to create a new update on a specific item or as a reply to another update.
+        Execute a mutation to create a new update on a specific item or as a reply to another update.
 
         For more information, visit https://developer.monday.com/api-reference/reference/updates#create-an-update
 
@@ -62,12 +77,14 @@ class UpdateResource(AsyncBaseResource):
             parent_id (ID): (Optional) The ID of the parent update to reply to.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = create_update_query(body=body, item_id=item_id, parent_id=parent_id, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = create_update_mutation(
+            body=body, item_id=item_id, parent_id=parent_id, with_complexity=with_complexity
+        )
+        return await self.client.execute(mutation)
 
     async def edit_update(self, update_id: ID, body: str, with_complexity: bool = False) -> dict:
         """
-        Edit the content of an update.
+        Execute a mutation to edit the content of an update.
         For more information, visit https://developer.monday.com/api-reference/reference/updates#edit-an-update
 
         Args:
@@ -78,12 +95,12 @@ class UpdateResource(AsyncBaseResource):
         Returns:
             dict: The JSON response from the GraphQL server.
         """
-        query = edit_update_query(update_id=update_id, body=body, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = edit_update_mutation(update_id=update_id, body=body, with_complexity=with_complexity)
+        return await self.client.execute(mutation)
 
     async def pin_update(self, update_id: ID, with_complexity: bool = False) -> dict:
         """
-        Execute a query to pin  an update to the top of the updates section of a specific item.
+        Execute a mutation to pin an update to the top of the updates section of a specific item.
         For more information, visit https://developer.monday.com/api-reference/reference/updates#pin-an-update
 
         Args:
@@ -93,12 +110,12 @@ class UpdateResource(AsyncBaseResource):
         Returns:
             dict: The JSON response from the GraphQL server.
         """
-        query = pin_update_query(update_id=update_id, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = pin_update_mutation(update_id=update_id, with_complexity=with_complexity)
+        return await self.client.execute(mutation)
 
     async def unpin_update(self, update_id: ID, with_complexity: bool = False) -> dict:
         """
-        Execute a query to unpin an update from the top of the updates section of a specific item.
+        Execute a mutation to unpin an update from the top of the updates section of a specific item.
         For more information, visit https://developer.monday.com/api-reference/reference/updates#unpin-an-update
 
         Args:
@@ -108,12 +125,12 @@ class UpdateResource(AsyncBaseResource):
         Returns:
             dict: The JSON response from the GraphQL server.
         """
-        query = unpin_update_query(update_id=update_id, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = unpin_update_mutation(update_id=update_id, with_complexity=with_complexity)
+        return await self.client.execute(mutation)
 
     async def like_update(self, update_id: ID, with_complexity: bool = False) -> dict:
         """
-        Execute a query to add a like to a specific update.
+        Execute a mutation to add a like to a specific update.
 
         For more information, visit https://developer.monday.com/api-reference/reference/updates#like-an-update
 
@@ -121,12 +138,12 @@ class UpdateResource(AsyncBaseResource):
             update_id (ID): The ID of the update to like.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = like_update_query(update_id=update_id, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = like_update_mutation(update_id=update_id, with_complexity=with_complexity)
+        return await self.client.execute(mutation)
 
     async def unlike_update(self, update_id: ID, with_complexity: bool = False) -> dict:
         """
-        Execute a query to remove a like from a specific update.
+        Execute a mutation to remove a like from a specific update.
         For more information, visit https://developer.monday.com/api-reference/reference/updates#unlike-an-update
 
         Args:
@@ -136,24 +153,24 @@ class UpdateResource(AsyncBaseResource):
         Returns:
             dict: The JSON response from the GraphQL server.
         """
-        query = unlike_update_query(update_id=update_id, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = unlike_update_mutation(update_id=update_id, with_complexity=with_complexity)
+        return await self.client.execute(mutation)
 
     async def delete_update(self, update_id: ID, with_complexity: bool = False) -> dict:
         """
-        Execute a query to remove an update. For more information, visit
+        Execute a mutation to remove an update. For more information, visit
         https://developer.monday.com/api-reference/reference/updates#delete-an-update
 
         Args:
             update_id (ID): The unique identifier of the update to delete.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = delete_update_query(update_id=update_id, with_complexity=with_complexity)
-        return await self.client.execute(query)
+        mutation = delete_update_mutation(update_id=update_id, with_complexity=with_complexity)
+        return await self.client.execute(mutation)
 
     async def add_file_to_update(self, update_id: ID, file: str, with_complexity: bool = False) -> dict:
         """
-        Execute a query to add a file to an update. For more information, visit
+        Execute a mutation to add a file to an update. For more information, visit
         https://developer.monday.com/api-reference/reference/assets-1#add-a-file-to-an-update
 
         Args:
@@ -161,5 +178,5 @@ class UpdateResource(AsyncBaseResource):
             file (str): The filepath to the file.
             with_complexity (bool): Set to True to return the query's complexity along with the results.
         """
-        query = add_file_to_update(update_id=update_id, with_complexity=with_complexity)
-        return await self.file_upload_client.execute(query, variables={"file": file})
+        mutation = add_file_to_update_mutation(update_id=update_id, with_complexity=with_complexity)
+        return await self.file_upload_client.execute(mutation, variables={"file": file})
