@@ -13,19 +13,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from monday_async.resources.api import APIResource
-from monday_async.resources.custom import CustomResource
-from monday_async.resources.webhooks import WebhooksResource
-from monday_async.resources.notifications import NotificationResource
-from monday_async.resources.users import UsersResource
-from monday_async.resources.teams import TeamsResource
-from monday_async.resources.workspaces import WorkspaceResource
-from monday_async.resources.folders import FolderResource
-from monday_async.resources.boards import BoardResource
-from monday_async.resources.tags import TagResource
-from monday_async.resources.columns import ColumnResource
-from monday_async.resources.groups import GroupResource
-from monday_async.resources.items import ItemResource
-from monday_async.resources.updates import UpdateResource
-from monday_async.resources.complexity import ComplexityResource
-from monday_async.resources.account import AccountResource
+from monday_async.resources.account import AccountResource as AccountResource
+from monday_async.resources.api import APIResource as APIResource
+from monday_async.resources.boards import BoardResource as BoardResource
+from monday_async.resources.columns import ColumnResource as ColumnResource
+from monday_async.resources.complexity import ComplexityResource as ComplexityResource
+from monday_async.resources.custom import CustomResource as CustomResource
+from monday_async.resources.folders import FolderResource as FolderResource
+from monday_async.resources.groups import GroupResource as GroupResource
+from monday_async.resources.items import ItemResource as ItemResource
+from monday_async.resources.notifications import NotificationResource as NotificationResource
+from monday_async.resources.tags import TagResource as TagResource
+from monday_async.resources.teams import TeamsResource as TeamsResource
+from monday_async.resources.updates import UpdateResource as UpdateResource
+from monday_async.resources.users import UsersResource as UsersResource
+from monday_async.resources.webhooks import WebhooksResource as WebhooksResource
+from monday_async.resources.workspaces import WorkspaceResource as WorkspaceResource
