@@ -15,91 +15,44 @@
 
 import pytest
 
-from monday_async.utils.queries.query_addons import (
-    add_complexity,
-    add_columns,
-    add_groups,
+from monday_async.graphql.addons import (
     add_column_values,
+    add_columns,
+    add_complexity,
+    add_groups,
     add_subitems,
-    add_updates
+    add_updates,
 )
 
 
 @pytest.mark.parametrize(
     "func, required_fields",
     [
+        (add_complexity, ["complexity", "before", "query", "after", "reset_in_x_seconds"]),
+        (add_columns, ["columns", "id", "title", "type", "settings_str"]),
+        (add_groups, ["groups", "id", "title", "color", "position"]),
         (
-                add_complexity,
-                [
-                    "complexity",
-                    "before",
-                    "query",
-                    "after",
-                    "reset_in_x_seconds"
-                ]
+            add_column_values,
+            [
+                "column_values",
+                "column",
+                "title",
+                "settings_str",
+                "type",
+                "value",
+                "text",
+                "... on BoardRelationValue",
+                "... on CheckboxValue",
+                "... on CountryValue",
+                "... on DateValue",
+                "... on LocationValue",
+                "... on MirrorValue",
+                "... on PeopleValue",
+            ],
         ),
-        (
-                add_columns,
-                [
-                    "columns",
-                    "id",
-                    "title",
-                    "type",
-                    "settings_str"
-                ]
-        ),
-        (
-                add_groups,
-                [
-                    "groups",
-                    "id",
-                    "title",
-                    "color",
-                    "position"
-                ]
-        ),
-        (
-                add_column_values,
-                [
-                    "column_values",
-                    "column",
-                    "title",
-                    "settings_str",
-                    "type",
-                    "value",
-                    "text",
-                    "... on BoardRelationValue",
-                    "... on CheckboxValue",
-                    "... on CountryValue",
-                    "... on DateValue",
-                    "... on LocationValue",
-                    "... on MirrorValue",
-                    "... on PeopleValue",
-                ]
-        ),
-        (
-                add_subitems,
-                [
-                    "subitems",
-                    "id",
-                    "name",
-                    "url",
-                    "state"
-                ]
-        ),
-        (
-                add_updates,
-                [
-                    "updates",
-                    "id",
-                    "text_body",
-                    "body",
-                    "creator_id",
-                    "assets",
-                    "replies"
-                ]
-        ),
-    ]
+        (add_subitems, ["subitems", "id", "name", "url", "state"]),
+        (add_updates, ["updates", "id", "text_body", "body", "creator_id", "assets", "replies"]),
+    ],
 )
 def test_query_addons_functions(func, required_fields):
     """
