@@ -18,7 +18,7 @@ from typing import Any
 
 import pytest
 
-from monday_async.utils.utils import monday_json_stringify, graphql_parse, format_param_value, format_dict_value
+from monday_async.core.helpers import format_dict_value, format_param_value, graphql_parse, monday_json_stringify
 
 
 class EnumForTesting(Enum):
@@ -33,23 +33,19 @@ class EnumForTesting(Enum):
         # Simple cases
         ({"label": "Done"}, '"{\\"label\\":\\"Done\\"}"'),
         ({"label": "Done", "status": "Completed"}, '"{\\"label\\":\\"Done\\",\\"status\\":\\"Completed\\"}"'),
-
         # Special characters and emojis
         ({"emoji": "😊🌟🎉"}, '"{\\"emoji\\":\\"😊🌟🎉\\"}"'),
         ({"symbols": "!@#$%^&*()_+"}, '"{\\"symbols\\":\\"!@#$%^&*()_+\\"}"'),
-
         # Non-Latin scripts
         ({"cyrillic": "Привет мир"}, '"{\\"cyrillic\\":\\"Привет мир\\"}"'),
         ({"japanese": "日本語"}, '"{\\"japanese\\":\\"日本語\\"}"'),
-
         # Nested structures
         (
-                {"nested": {"key": "Value with 😊", "num": 42}},
-                '"{\\"nested\\":{\\"key\\":\\"Value with 😊\\",\\"num\\":42}}"'
+            {"nested": {"key": "Value with 😊", "num": 42}},
+            '"{\\"nested\\":{\\"key\\":\\"Value with 😊\\",\\"num\\":42}}"',
         ),
-
         # Edge cases
-        (None, 'null'),
+        (None, "null"),
         ({}, '"{}"'),
         ({"empty": ""}, '"{\\"empty\\":\\"\\"}"'),
         ({"numbers": 123}, '"{\\"numbers\\":123}"'),
@@ -67,8 +63,8 @@ class EnumForTesting(Enum):
         "empty_dict",
         "empty_string_value",
         "numeric_value",
-        "accented_chars"
-    ]
+        "accented_chars",
+    ],
 )
 def test_monday_json_stringify(value, expected_output):
     assert monday_json_stringify(value) == expected_output
@@ -79,14 +75,20 @@ def test_monday_json_stringify(value, expected_output):
     "query, expected",
     [
         ("query { items { id } }", "{\n  items {\n    id\n  }\n}"),
-        ("query($id: ID!) {items(ids: [$id]) { name } }",
-         "query ($id: ID!) {\n  items(ids: [$id]) {\n    name\n  }\n}"),
-        ("mutation { create_item (board_id: 123, item_name: \"Test\") { id } }",
-         "mutation {\n  create_item(board_id: 123, item_name: \"Test\") {\n    id\n  }\n}"),
-        ("\n\nquery\n($id: ID!\n\t) \t\t\t{items(ids: [$id]) { name \tid} }",
-         "query ($id: ID!) {\n  items(ids: [$id]) {\n    name\n    id\n  }\n}")
+        (
+            "query($id: ID!) {items(ids: [$id]) { name } }",
+            "query ($id: ID!) {\n  items(ids: [$id]) {\n    name\n  }\n}",
+        ),
+        (
+            'mutation { create_item (board_id: 123, item_name: "Test") { id } }',
+            'mutation {\n  create_item(board_id: 123, item_name: "Test") {\n    id\n  }\n}',
+        ),
+        (
+            "\n\nquery\n($id: ID!\n\t) \t\t\t{items(ids: [$id]) { name \tid} }",
+            "query ($id: ID!) {\n  items(ids: [$id]) {\n    name\n    id\n  }\n}",
+        ),
     ],
-    ids=["simple_query", "query_with_variables", "mutation", "normalize_whitespace"]
+    ids=["simple_query", "query_with_variables", "mutation", "normalize_whitespace"],
 )
 def test_graphql_parse_valid_queries(query, expected):
     """Test that valid GraphQL queries are properly formatted"""
@@ -99,23 +101,14 @@ def test_graphql_parse_valid_queries(query, expected):
     [
         (EnumForTesting.PENDING, "pending"),
         (123, "123"),
-        ("test", "\"test\""),
+        ("test", '"test"'),
         (True, "true"),
         (None, "null"),
-        (["a", 1], "[\"a\", 1]"),
-        ({"key": "value"}, "{\"key\": \"value\"}"),
-        (3.14, "3.14")
+        (["a", 1], '["a", 1]'),
+        ({"key": "value"}, '{"key": "value"}'),
+        (3.14, "3.14"),
     ],
-    ids=[
-        "enum",
-        "int",
-        "str",
-        "bool",
-        "none",
-        "list",
-        "dict",
-        "float"
-    ]
+    ids=["enum", "int", "str", "bool", "none", "list", "dict", "float"],
 )
 def test_format_param_value(value: Any, expected: str):
     """Test various value types formatting"""
@@ -124,24 +117,18 @@ def test_format_param_value(value: Any, expected: str):
 
 # Test cases for format_dict_value
 
+
 @pytest.mark.parametrize(
     "input_dict,expected",
     [
         ({"status": EnumForTesting.PENDING, "count": 5}, "{status: pending, count: 5}"),
         ({}, "{}"),
-        ({"name": "Alice", "age": 30}, "{name: \"Alice\", age: 30}"),
-        ({"nested": {"key": "value"}}, "{nested: {\"key\": \"value\"}}"),
-        ({"special_chars": "áéíóú"}, "{special_chars: \"áéíóú\"}"),
-        ({"emoji": "😊"}, "{emoji: \"😊\"}"),
+        ({"name": "Alice", "age": 30}, '{name: "Alice", age: 30}'),
+        ({"nested": {"key": "value"}}, '{nested: {"key": "value"}}'),
+        ({"special_chars": "áéíóú"}, '{special_chars: "áéíóú"}'),
+        ({"emoji": "😊"}, '{emoji: "😊"}'),
     ],
-    ids=[
-        "enum_value",
-        "empty_dict",
-        "simple_key_value",
-        "nested_dict",
-        "special_chars",
-        "emoji"
-    ]
+    ids=["enum_value", "empty_dict", "simple_key_value", "nested_dict", "special_chars", "emoji"],
 )
 def test_format_dict_value(input_dict: dict, expected: str):
     """Test dictionary formatting with various value types"""
